chore: remove hardened images
Having "hardened" images makes it so out regular ones look like they dont have any security or anything, and most of the tweaks from the hardened ones break many of the tweaks that this distro makes. Also it is a maintenance burden to worry about these images and everything around them.